+++ frontend/hooks/work/useWorkOriginals.ts
@@ -1,17 +1,17 @@
 import { Work, Original } from "model"
 import { useState, useEffect } from "react"
-import { WorkRepository } from "repository/work"
+import { OriginalRepository } from "repository"
 
 export const useWorkOriginals = (work: Work): { loading: boolean, originals: Original[] } => {
   const [originals, setOriginals] = useState<Original[]>([])
   const [loading, setLoading] = useState(true)
-  const workRepo = new WorkRepository()
+  const originalRepo = new OriginalRepository()
 
   useEffect(() => {
     let cancel = false
     const retrive = async () => {
       setLoading(true)
-      const originals = await workRepo.fetchOriginals(work.id)
+      const originals = await originalRepo.fetchOriginals(work.id)
       if (!cancel) {
         setOriginals(originals)
         setLoading(false)
